+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/generators/HttpProtocolGenerator.kt
@@ -12,11 +12,14 @@ import software.amazon.smithy.model.shapes.ServiceShape
 import software.amazon.smithy.model.shapes.ShapeId
 import software.amazon.smithy.model.shapes.StructureShape
 import software.amazon.smithy.rust.codegen.rustlang.Attribute
+import software.amazon.smithy.rust.codegen.rustlang.CargoDependency
 import software.amazon.smithy.rust.codegen.rustlang.RustWriter
+import software.amazon.smithy.rust.codegen.rustlang.asType
 import software.amazon.smithy.rust.codegen.rustlang.docs
 import software.amazon.smithy.rust.codegen.rustlang.documentShape
 import software.amazon.smithy.rust.codegen.rustlang.rust
 import software.amazon.smithy.rust.codegen.rustlang.rustBlock
+import software.amazon.smithy.rust.codegen.rustlang.rustBlockTemplate
 import software.amazon.smithy.rust.codegen.rustlang.rustTemplate
 import software.amazon.smithy.rust.codegen.rustlang.withBlock
 import software.amazon.smithy.rust.codegen.smithy.RuntimeConfig
@@ -24,6 +27,7 @@ import software.amazon.smithy.rust.codegen.smithy.RuntimeType
 import software.amazon.smithy.rust.codegen.smithy.RustSymbolProvider
 import software.amazon.smithy.rust.codegen.smithy.customize.OperationCustomization
 import software.amazon.smithy.rust.codegen.smithy.customize.OperationSection
+import software.amazon.smithy.rust.codegen.smithy.customize.writeCustomizations
 import software.amazon.smithy.rust.codegen.smithy.letIf
 import software.amazon.smithy.rust.codegen.util.dq
 import software.amazon.smithy.rust.codegen.util.getTrait
@@ -52,12 +56,46 @@ interface ProtocolGeneratorFactory<out T : HttpProtocolGenerator> {
  * Abstract class providing scaffolding for HTTP based protocols that must build an HTTP request (headers / URL) and
  * a body.
  */
-abstract class HttpProtocolGenerator(
-    private val protocolConfig: ProtocolConfig
-) {
+abstract class HttpProtocolGenerator(protocolConfig: ProtocolConfig) {
+    private val runtimeConfig = protocolConfig.runtimeConfig
     private val symbolProvider = protocolConfig.symbolProvider
     private val model = protocolConfig.model
-    private val buildErrorT = protocolConfig.runtimeConfig.operationBuildError()
+
+    private val sdkId =
+        protocolConfig.serviceShape.getTrait<ServiceTrait>()?.sdkId?.toLowerCase()?.replace(" ", "")
+            ?: protocolConfig.serviceShape.id.getName(protocolConfig.serviceShape)
+
+    private val codegenScope = arrayOf(
+        "HttpRequestBuilder" to RuntimeType.HttpRequestBuilder,
+        "OpBuildError" to protocolConfig.runtimeConfig.operationBuildError(),
+        "Request" to RuntimeType.Http("request::Request"),
+        "RequestBuilder" to RuntimeType.HttpRequestBuilder,
+        "SdkBody" to RuntimeType.sdkBody(protocolConfig.runtimeConfig),
+        "config" to RuntimeType.Config,
+        "header_util" to CargoDependency.SmithyHttp(protocolConfig.runtimeConfig).asType().member("header"),
+        "http" to RuntimeType.http,
+        "operation" to RuntimeType.operationModule(runtimeConfig),
+    )
+
+    data class BodyMetadata(val takesOwnership: Boolean)
+
+    abstract fun RustWriter.body(self: String, operationShape: OperationShape): BodyMetadata
+
+    abstract fun traitImplementations(operationWriter: RustWriter, operationShape: OperationShape)
+
+    /** Write code into the impl block for [operationShape] */
+    open fun operationImplBlock(implBlockWriter: RustWriter, operationShape: OperationShape) {}
+
+    /**
+     * Add necessary methods to the impl block for the input shape.
+     *
+     * Your implementation MUST call [generateRequestBuilderBase] to create the public method.
+     */
+    abstract fun toHttpRequestImpl(
+        implBlockWriter: RustWriter,
+        operationShape: OperationShape,
+        inputShape: StructureShape
+    )
 
     fun renderOperation(
         operationWriter: RustWriter,
@@ -66,9 +104,6 @@ abstract class HttpProtocolGenerator(
         customizations: List<OperationCustomization>
     ) {
         val inputShape = operationShape.inputShape(model)
-        val sdkId =
-            protocolConfig.serviceShape.getTrait<ServiceTrait>()?.sdkId?.toLowerCase()?.replace(" ", "")
-                ?: protocolConfig.serviceShape.id.getName(protocolConfig.serviceShape)
         val builderGenerator = BuilderGenerator(model, symbolProvider, operationShape.inputShape(model))
         builderGenerator.render(inputWriter)
 
@@ -86,30 +121,35 @@ abstract class HttpProtocolGenerator(
         )
 
         // impl OperationInputShape { ... }
+        val operationName = symbolProvider.toSymbol(operationShape).name
         inputWriter.implBlock(inputShape, symbolProvider) {
-            buildOperation(this, operationShape, customizations, sdkId)
+            writeCustomizations(customizations, OperationSection.InputImpl(operationShape, inputShape))
+            generateMakeOperation(this, operationShape, operationName, customizations)
             toHttpRequestImpl(this, operationShape, inputShape)
-            rustBlock(
-                "fn assemble(mut builder: #1T, body: #3T) -> #2T<#3T>",
-                RuntimeType.HttpRequestBuilder,
-                RuntimeType.Http("request::Request"),
-                RuntimeType.sdkBody(protocolConfig.runtimeConfig)
+            rustBlockTemplate(
+                "fn assemble(mut builder: #{RequestBuilder}, body: #{SdkBody}) -> #{Request}<#{SdkBody}>",
+                *codegenScope
             ) {
                 rustTemplate(
                     """
                     if let Some(content_length) = body.content_length() {
-                        builder = builder.header(#{http}::header::CONTENT_LENGTH, content_length)
+                        builder = #{header_util}::set_header_if_absent(
+                                    builder,
+                                    #{http}::header::CONTENT_LENGTH,
+                                    content_length
+                        );
                     }
                     builder.body(body).expect("should be valid request")
-                """,
-                    "http" to RuntimeType.http
+                    """,
+                    *codegenScope
                 )
             }
 
             // pub fn builder() -> ... { }
             builderGenerator.renderConvenienceMethod(this)
         }
-        val operationName = symbolProvider.toSymbol(operationShape).name
+
+        // pub struct Operation { ... }
         operationWriter.documentShape(operationShape, model)
         Attribute.Derives(setOf(RuntimeType.Clone, RuntimeType.Default, RuntimeType.Debug)).render(operationWriter)
         operationWriter.rustBlock("pub struct $operationName") {
@@ -124,34 +164,29 @@ abstract class HttpProtocolGenerator(
                 rust("Self { _private: () }")
             }
 
-            customizations.forEach { customization -> customization.section(OperationSection.OperationImplBlock)(this) }
+            writeCustomizations(customizations, OperationSection.OperationImplBlock)
         }
         traitImplementations(operationWriter, operationShape)
     }
 
-    protected fun httpBuilderFun(implBlockWriter: RustWriter, f: RustWriter.() -> Unit) {
+    protected fun generateRequestBuilderBase(implBlockWriter: RustWriter, f: RustWriter.() -> Unit) {
         Attribute.Custom("allow(clippy::unnecessary_wraps)").render(implBlockWriter)
-        implBlockWriter.rustBlock(
-            "fn request_builder_base(&self) -> std::result::Result<#T, #T>",
-            RuntimeType.HttpRequestBuilder, buildErrorT
+        implBlockWriter.rustBlockTemplate(
+            "fn request_builder_base(&self) -> std::result::Result<#{HttpRequestBuilder}, #{OpBuildError}>",
+            *codegenScope,
         ) {
             f(this)
         }
     }
 
-    data class BodyMetadata(val takesOwnership: Boolean)
-
-    abstract fun RustWriter.body(self: String, operationShape: OperationShape): BodyMetadata
-
     private fun buildOperationType(
         writer: RustWriter,
         shape: OperationShape,
-        features: List<OperationCustomization>,
+        customizations: List<OperationCustomization>,
     ): String {
-        val runtimeConfig = protocolConfig.runtimeConfig
         val operationT = RuntimeType.operation(runtimeConfig)
         val output = buildOperationTypeOutput(writer, shape)
-        val retry = buildOperationTypeRetry(writer, features)
+        val retry = buildOperationTypeRetry(writer, customizations)
 
         return with(writer) { "${format(operationT)}<$output, $retry>" }
     }
@@ -159,80 +194,55 @@ abstract class HttpProtocolGenerator(
     private fun buildOperationTypeOutput(writer: RustWriter, shape: OperationShape): String =
         writer.format(symbolProvider.toSymbol(shape))
 
-    private fun buildOperationTypeRetry(writer: RustWriter, features: List<OperationCustomization>): String =
-        features.mapNotNull { it.retryType() }.firstOrNull()?.let { writer.format(it) } ?: "()"
+    private fun buildOperationTypeRetry(writer: RustWriter, customizations: List<OperationCustomization>): String =
+        customizations.mapNotNull { it.retryType() }.firstOrNull()?.let { writer.format(it) } ?: "()"
 
-    private fun buildOperation(
+    private fun generateMakeOperation(
         implBlockWriter: RustWriter,
         shape: OperationShape,
-        features: List<OperationCustomization>,
-        sdkId: String
+        operationName: String,
+        customizations: List<OperationCustomization>
     ) {
-        val runtimeConfig = protocolConfig.runtimeConfig
-        val outputSymbol = symbolProvider.toSymbol(shape)
-        val operationModule = RuntimeType.operationModule(runtimeConfig)
-        val sdkBody = RuntimeType.sdkBody(runtimeConfig)
-
-        val baseReturnType = buildOperationType(implBlockWriter, shape, features)
+        val baseReturnType = buildOperationType(implBlockWriter, shape, customizations)
         val returnType = "std::result::Result<$baseReturnType, ${implBlockWriter.format(runtimeConfig.operationBuildError())}>"
+        val outputSymbol = symbolProvider.toSymbol(shape)
+
+        val bodyMetadata = RustWriter.root().body("self", shape)
+        val mut = customizations.any { it.mutSelf() }
+        val consumes = customizations.any { it.consumesSelf() } || bodyMetadata.takesOwnership
+        val self = "self".letIf(mut) { "mut $it" }.letIf(!consumes) { "&$it" }
 
         implBlockWriter.docs("Consumes the builder and constructs an Operation<#D>", outputSymbol)
-        // For codegen simplicity, allow `let x = ...; x`
-        implBlockWriter.rust("##[allow(clippy::let_and_return)]")
-        val bodyMetadata = RustWriter.root().body("self", shape)
-        val mut = features.any { it.mutSelf() }
-        val consumes = features.any { it.consumesSelf() } || bodyMetadata.takesOwnership
-        val self = "self".letIf(mut) { "mut $it" }.letIf(!consumes) { "&$it" }
-        implBlockWriter.rustBlock(
-            "pub fn make_operation($self, _config: &#T::Config) -> $returnType",
-            RuntimeType.Config
+        implBlockWriter.rust("##[allow(clippy::let_and_return)]") // For codegen simplicity, allow `let x = ...; x`
+        implBlockWriter.rustBlockTemplate(
+            "pub fn make_operation($self, _config: &#{config}::Config) -> $returnType",
+            *codegenScope
         ) {
-            withBlock("Ok({", "})") {
-                features.forEach { it.section(OperationSection.MutateInput("self", "_config"))(this) }
-                rust("let properties = smithy_http::property_bag::SharedPropertyBag::new();")
-                rust("let request = self.request_builder_base()?;")
-                withBlock("let body =", ";") {
-                    body("self", shape)
-                }
-                rust("let request = Self::assemble(request, body);")
-                rust(
-                    """
-                    ##[allow(unused_mut)]
-                    let mut request = #T::Request::from_parts(request.map(#T::from), properties);
-                """,
-                    operationModule, sdkBody
-                )
-                features.forEach { it.section(OperationSection.MutateRequest("request", "_config"))(this) }
-                rust(
-                    """
-                    let op = #1T::Operation::new(
-                        request,
-                        #2T::new()
-                    ).with_metadata(#1T::Metadata::new(${
-                    shape.id.getName(protocolConfig.serviceShape).dq()
-                    }, ${sdkId.dq()}));
-                """,
-                    operationModule, symbolProvider.toSymbol(shape)
-                )
-                features.forEach { it.section(OperationSection.FinalizeOperation("op", "_config"))(this) }
-                rust("op")
+            writeCustomizations(customizations, OperationSection.MutateInput("self", "_config"))
+            rust("let properties = smithy_http::property_bag::SharedPropertyBag::new();")
+            rust("let request = self.request_builder_base()?;")
+            withBlock("let body =", ";") {
+                body("self", shape)
             }
+            rust("let request = Self::assemble(request, body);")
+            rustTemplate(
+                """
+                ##[allow(unused_mut)]
+                let mut request = #{operation}::Request::from_parts(request.map(#{SdkBody}::from), properties);
+                """,
+                *codegenScope
+            )
+            writeCustomizations(customizations, OperationSection.MutateRequest("request", "_config"))
+            rustTemplate(
+                """
+                let op = #{operation}::Operation::new(request, #{OperationType}::new())
+                    .with_metadata(#{operation}::Metadata::new(${operationName.dq()}, ${sdkId.dq()}));
+                """,
+                *codegenScope,
+                "OperationType" to symbolProvider.toSymbol(shape)
+            )
+            writeCustomizations(customizations, OperationSection.FinalizeOperation("op", "_config"))
+            rust("Ok(op)")
         }
     }
-
-    abstract fun traitImplementations(operationWriter: RustWriter, operationShape: OperationShape)
-
-    /** Write code into the impl block for [operationShape] */
-    open fun operationImplBlock(implBlockWriter: RustWriter, operationShape: OperationShape) {}
-
-    /**
-     * Add necessary methods to the impl block for the input shape.
-     *
-     * Your implementation MUST call [httpBuilderFun] to create the public method.
-     */
-    abstract fun toHttpRequestImpl(
-        implBlockWriter: RustWriter,
-        operationShape: OperationShape,
-        inputShape: StructureShape
-    )
 }